edumeal-recommender/requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
pandas
numpy
faiss-cpu
sentence-transformers
scikit-learn
python-dotenv
psycopg[binary]
SQLAlchemy>=2
pgvector
networkx
xgboost

